Randomizer fixed?
Added some janky if statements to make the killer randomizer not show duplicates. There is def a more efficient way of doing it, but it works for now so ill keep it like that. Also added, oppressor and scouragehookgiftofpain pngs since I seemed to have forgotten those. As well as downloaded a more HD version of Deadlock png.