Survivor Perks finished plus media queries for some sizes
I will most likely come back and work on this more because theres a lot to do and I cant be bothered resizing every single element. But, the main goal of this website was for me to create a Perk randomizer for DBD and I completed that! I just wish I started off with a better sizing solution at the beginning but I was extremely new at the time. I hope to return and make this better with more features.